Fix: If the gallery selection is cancelled, it should not attempt to deleted the selected file unless its from camera

+++ android/src/main/kotlin/ts/tally/document_scanner/fallback/utils/DocumentProviderUtil.kt
@@ -60,7 +60,9 @@ class DocumentProviderUtil(
             }
             Activity.RESULT_CANCELED -> {
                 // delete the photo since the user didn't finish taking the photo
-                File(documentPath).delete()
+                if (documentPath != null) {
+                    File(documentPath).delete()
+                }
                 onCancelDocumentSelect()
             }
         }
